+++ crates/emmylua_code_analysis/src/db_index/signature/signature.rs
@@ -11,11 +11,11 @@ use crate::{
     FileId,
     db_index::{LuaFunctionType, LuaType},
 };
-use crate::{SemanticModel, VariadicType};
+use crate::{LuaAttributeUse, SemanticModel, VariadicType};
 
 #[derive(Debug)]
 pub struct LuaSignature {
-    pub generic_params: Vec<(String, Option<LuaType>)>,
+    pub generic_params: Vec<Arc<LuaGenericParamInfo>>,
     pub overloads: Vec<Arc<LuaFunctionType>>,
     pub param_docs: HashMap<usize, LuaDocParamInfo>,
     pub params: Vec<String>,
@@ -183,6 +183,16 @@ pub struct LuaDocParamInfo {
     pub type_ref: LuaType,
     pub nullable: bool,
     pub description: Option<String>,
+    pub attributes: Option<Vec<LuaAttributeUse>>,
+}
+
+impl LuaDocParamInfo {
+    pub fn get_attribute_by_name(&self, name: &str) -> Option<&LuaAttributeUse> {
+        self.attributes
+            .iter()
+            .flatten()
+            .find(|attr| attr.id.get_name() == name)
+    }
 }
 
 #[derive(Debug)]
@@ -190,6 +200,7 @@ pub struct LuaDocReturnInfo {
     pub name: Option<String>,
     pub type_ref: LuaType,
     pub description: Option<String>,
+    pub attributes: Option<Vec<LuaAttributeUse>>,
 }
 
 #[derive(Debug, Hash, Eq, PartialEq, Clone, Copy)]
@@ -279,4 +290,25 @@ pub enum SignatureReturnStatus {
     UnResolve,
     DocResolve,
     InferResolve,
+}
+
+#[derive(Debug, Clone)]
+pub struct LuaGenericParamInfo {
+    pub name: String,
+    pub type_constraint: Option<LuaType>,
+    pub attributes: Option<Vec<LuaAttributeUse>>,
+}
+
+impl LuaGenericParamInfo {
+    pub fn new(
+        name: String,
+        type_constraint: Option<LuaType>,
+        attributes: Option<Vec<LuaAttributeUse>>,
+    ) -> Self {
+        Self {
+            name,
+            type_constraint,
+            attributes,
+        }
+    }
 }